Move behat tests to the new folder structure
