--- behat/api.yml
@@ -1,34 +1,0 @@
-api:
-  extensions:
-    Behat\Symfony2Extension:
-      kernel:
-        class: CodelyTv\Api\ApiKernel
-    Behat\MinkExtension:
-      sessions:
-        my_session:
-          symfony2: ~
-
-  suites:
-    status:
-      paths: [ applications/api/features/status ]
-      contexts:
-        - CodelyTv\Test\EntryPoint\ApiFeatureContext:
-            connections: '@codely.infrastructure.database_connections'
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiRequestContext
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiResponseContext
-    video:
-      paths: [ applications/api/features/video ]
-      contexts:
-        - CodelyTv\Test\EntryPoint\ApiFeatureContext:
-            connections: '@codely.infrastructure.database_connections'
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiRequestContext
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiResponseContext
-    user:
-      paths: [ applications/api/features/user ]
-      contexts:
-        - CodelyTv\Test\EntryPoint\ApiFeatureContext:
-            connections: '@codely.infrastructure.database_connections'
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiRequestContext
-        - CodelyTv\Test\EntryPoint\ApiContext\ApiResponseContext
-        - CodelyTv\Test\Context\Video\Module\User\UserModuleBehatContext:
-            repository: '@codely.video.user.repository'
