+++ tests/Context/Video/Module/User/Domain/TotalVideosCreatedStub.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace CodelyTv\Test\Context\Video\Module\User\Domain;
 
 use CodelyTv\Context\Video\Module\User\Domain\TotalVideosCreated;
-use CodelyTv\Test\Infrastructure\Stub\NumberStub;
+use CodelyTv\Test\Shared\Domain\NumberStub;
 
 final class TotalVideosCreatedStub
 {
