+++ tests/Context/Video/Module/User/Domain/UserNameStub.php
@@ -5,7 +5,7 @@ declare(strict_types = 1);
 namespace CodelyTv\Test\Context\Video\Module\User\Domain;
 
 use CodelyTv\Context\Video\Module\User\Domain\UserName;
-use CodelyTv\Test\Infrastructure\Stub\WordStub;
+use CodelyTv\Test\Shared\Domain\WordStub;
 
 final class UserNameStub
 {
