+++ tests/Context/Video/Module/Video/Application/Create/CreateVideoCommandStub.php
@@ -13,7 +13,7 @@ use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoIdStub;
 use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoTitleStub;
 use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoTypeStub;
 use CodelyTv\Test\Context\Video\Module\Video\Domain\VideoUrlStub;
-use CodelyTv\Test\Infrastructure\Stub\UuidStub;
+use CodelyTv\Test\Shared\Domain\UuidStub;
 use CodelyTv\Types\ValueObject\Uuid;
 
 final class CreateVideoCommandStub
