+++ tests/Context/Video/Module/Video/Domain/VideoIdStub.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace CodelyTv\Test\Context\Video\Module\Video\Domain;
 
 use CodelyTv\Context\Video\Module\Video\Domain\VideoId;
-use CodelyTv\Test\Infrastructure\Stub\UuidStub;
+use CodelyTv\Test\Shared\Domain\UuidStub;
 
 final class VideoIdStub
 {
