+++ tests/Context/Video/Module/VideoComment/Domain/VideoCommentIdStub.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace CodelyTv\Test\Context\Video\Module\VideoComment\Domain;
 
 use CodelyTv\Context\Video\Module\VideoComment\Domain\VideoCommentId;
-use CodelyTv\Test\Infrastructure\Stub\UuidStub;
+use CodelyTv\Test\Shared\Domain\UuidStub;
 
 final class VideoCommentIdStub
 {
