+++ tests/EntryPoint/applications/features/bootstrap_check.feature
