+++ tests/Shared/Domain/DuplicatorStub.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CodelyTv\Test\Infrastructure\Stub;
+namespace CodelyTv\Test\Shared\Domain;
 
 use ReflectionObject;
 use ReflectionProperty;
